replaced 'set-name' bubble handler with on*
